--- iceoryx2-bb/testing-nostd/BUILD.bazel
@@ -1,30 +1,0 @@
-# Copyright (c) 2026 Contributors to the Eclipse Foundation
-#
-# See the NOTICE file(s) distributed with this work for additional
-# information regarding copyright ownership.
-#
-# This program and the accompanying materials are made available under the
-# terms of the Apache Software License 2.0 which is available at
-# https://www.apache.org/licenses/LICENSE-2.0, or the MIT license
-# which is available at https://opensource.org/licenses/MIT.
-#
-# SPDX-License-Identifier: Apache-2.0 OR MIT
-
-package(default_visibility = ["//visibility:public"])
-
-load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test_suite")
-
-filegroup(
-    name = "all_srcs",
-    srcs = glob(["**"]),
-)
-
-rust_library(
-    name = "iceoryx2-bb-testing-nostd",
-    srcs = glob(["src/**/*.rs"]),
-    deps = [
-        "//iceoryx2-pal/print:iceoryx2-pal-print",
-        "//iceoryx2-pal/posix:iceoryx2-pal-posix",
-        "@crate_index//:inventory",
-    ],
-)
